refactor(plugins): builtin 插件源码搬进根 plugins/ 目录(独立 Go 模块)
所有插件统一住 plugins/:feishu(飞书管家)、reviewmesh(智能互审)的 Go
源码从 cli/ttmux-cli-go/internal/plugin/builtin/ 迁出,成为独立模块
roam-plugins,经双向 replace 与 cli 模块互联——运行时不变,仍编译进
ttmux 二进制(builtin、开箱即用)。

- plugin SDK 从 internal/plugin/sdk 移到公开路径 pkg/plugin/sdk:插件只依赖
  公开 SDK,不再触碰宿主 internal(也是第三方 Go 插件未来的依赖面)
- cli 侧仅剩 builtin.go 注册表(manifest 声明 + Activate 接线)
- 质量门禁 Go tests 增加 plugins 模块;plugins/README.md 改写为两种形态说明
  (builtin Go 源码 vs roam-plugin.json 插件包)
