+++ plugins/reviewmesh/reviewmesh.go
@@ -17,7 +17,7 @@ import (
 	"strings"
 	"time"
 
-	"ttmux-cli-go/internal/plugin/sdk"
+	"ttmux-cli-go/pkg/plugin/sdk"
 )
 
 const (
